+++ lua/lspinstall/servers/html.lua
@@ -1,5 +1,13 @@
-return {
-  install_script = [[
+local lsp_util = require"lspinstall/util"
+
+local script_to_use = nil
+local cmd_to_use = ""
+
+if lsp_util.is_windows() then
+  --TODO somebody implement this if possible for windows
+else
+  cmd_to_use = { "node", "./vscode-html/html-language-features/server/dist/node/htmlServerMain.js", "--stdio" }
+  script_to_use  = [[
   curl -L -o vscode.tar.gz https://update.code.visualstudio.com/latest/linux-x64/stable
   rm -rf vscode
   mkdir vscode
@@ -12,9 +20,13 @@ return {
   cp -r vscode/resources/app/extensions/html-language-features vscode-html
 
   rm -rf vscode
-  ]],
+  ]]
+end
+
+return {
+  install_script = script_to_use,
   default_config = {
-    cmd = { "node", "./vscode-html/html-language-features/server/dist/node/htmlServerMain.js", "--stdio" },
+    cmd = cmd_to_use ,
     filetypes = {
       -- html
       'aspnetcorerazor',
